Refacto: extract types in dedicated files Fix: loss function is VAE
