+++ encyclopedia_vae/losses/losses.py
@@ -1,16 +1,11 @@
-from typing import TypedDict
-
 import torch
+from rich import print as pprint
 from torch.nn import functional as F
 
-
-class LossReturn(TypedDict):
-    loss: torch.tensor
-    reconstruction_loss: torch.tensor
-    kld_loss: torch.tensor
+from encyclopedia_vae.types import ForwardReturn, LossReturn
 
 
-def loss_function(self, input, recons, mu, log_var, kld_weight, **kwargs) -> LossReturn:
+def loss_function(output_model: ForwardReturn, kld_weight, **kwargs) -> LossReturn:
     """
     Computes the VAE loss function.
     KL(N(\mu, \sigma), N(0, 1)) = \log \frac{1}{\sigma} + \frac{\sigma^2 + \mu^2}{2} - \frac{1}{2}
@@ -20,10 +15,13 @@ def loss_function(self, input, recons, mu, log_var, kld_weight, **kwargs) -> Los
     """
 
     # kld_weight = kwargs["M_N"]  # Account for the minibatch samples from the dataset
-    recons_loss = F.mse_loss(recons, input)
+    recons_loss = F.mse_loss(output_model["output"], output_model["input"])
+    pprint(f"In loss function, MSE is {recons_loss}")
 
-    kld_loss = compute_kld_loss(mu, log_var)
-
+    kld_loss = compute_kld_loss(
+        output_model["latents"]["mu"], output_model["latents"]["log_var"]
+    )
+    pprint(f"In loss function, KLD is {kld_loss}")
     loss = recons_loss + kld_weight * kld_loss
     return LossReturn(loss=loss, reconstruction_loss=recons_loss, kld_loss=-kld_loss)
